+++ packages/analytics-node/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.51](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.50...@amplitude/analytics-node@1.5.51) (2026-03-26)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
 ## [1.5.50](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.49...@amplitude/analytics-node@1.5.50) (2026-03-23)
 
 **Note:** Version bump only for package @amplitude/analytics-node