+++ packages/plugin-network-capture-browser/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.11](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-network-capture-browser@1.9.10...@amplitude/plugin-network-capture-browser@1.9.11) (2026-03-26)
+
+**Note:** Version bump only for package @amplitude/plugin-network-capture-browser
+
+
+
+
+
 ## [1.9.10](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-network-capture-browser@1.9.9...@amplitude/plugin-network-capture-browser@1.9.10) (2026-03-23)
 
 **Note:** Version bump only for package @amplitude/plugin-network-capture-browser